feat: Retrieved guest houses from db, added role fields to models and restructure the web dirs
