+++ apps/server/.gitignore
@@ -51,3 +51,6 @@ lerna-debug.log*
 # typescript
 *.tsbuildinfo
 next-env.d.ts
+
+# Prisma generated client
+/prisma/src/generated
