--- apps/server/prisma/index.ts
@@ -1,5 +1,0 @@
-import { PrismaClient } from "./src/generated/client";
-
-const prisma = new PrismaClient();
-
-export default prisma;
