+++ apps/server/prisma/schema/schema.prisma
@@ -8,3 +8,5 @@ datasource db {
   provider = "postgres"
   url      = env("DATABASE_URL")
 }
+
+
