+++ apps/server/src/lib/prisma.ts
@@ -1,0 +1,5 @@
+import { PrismaClient } from "../../prisma/src/generated/client";
+
+const prisma = new PrismaClient();
+
+export default prisma;
