+++ apps/web/package.json
@@ -29,6 +29,7 @@
 		"@tanstack/react-form": "^1.12.3",
 		"@tanstack/react-query": "^5.80.5",
 		"@tanstack/react-table": "^8.21.3",
+		"axios": "^1.11.0",
 		"better-auth": "^1.3.4",
 		"class-variance-authority": "^0.7.1",
 		"clsx": "^2.1.1",
