+++ apps/web/src/app/(auth)/layout.tsx
@@ -1,0 +1,36 @@
+import { Geist, Geist_Mono } from "next/font/google";
+import "../../index.css";
+import type { Metadata } from "next";
+
+const geistSans = Geist({
+  variable: "--font-geist-sans",
+  subsets: ["latin"],
+});
+
+const geistMono = Geist_Mono({
+  variable: "--font-geist-mono",
+  subsets: ["latin"],
+});
+
+export const metadata: Metadata = {
+  title: "Sign In | Adminify",
+  description: "Sign In to Adminify",
+};
+
+export default function AuthLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="en" suppressHydrationWarning>
+      <body
+        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
+      >
+        <div className="flex min-h-screen items-center justify-center">
+          {children}
+        </div>
+      </body>
+    </html>
+  );
+}
