--- apps/web/src/app/dashboard/page.tsx
@@ -1,26 +1,0 @@
-"use client";
-import { authClient } from "@/lib/auth-client";
-import { useRouter } from "next/navigation";
-import { useEffect } from "react";
-
-export default function Dashboard() {
-	const router = useRouter();
-	const { data: session, isPending } = authClient.useSession();
-
-	useEffect(() => {
-		if (!session && !isPending) {
-			router.push("/login");
-		}
-	}, [session, isPending]);
-
-	if (isPending) {
-		return <div>Loading...</div>;
-	}
-
-	return (
-		<div>
-			<h1>Dashboard</h1>
-			<p>Welcome {session?.user.name}</p>
-		</div>
-	);
-}
