--- apps/web/src/app/page.tsx
@@ -1,14 +1,0 @@
-import { SectionCards } from "@/components/section-cards";
-
-export default function Page({
-  children,
-}: Readonly<{ children: React.ReactNode }>) {
-  return (
-    <div className="@container/main flex flex-1 flex-col gap-2">
-      <div className="flex flex-col gap-4 py-4 md:gap-6 md:py-6">
-        <SectionCards />
-        {children}
-      </div>
-    </div>
-  );
-}
