+++ apps/web/src/components/sign-up-form.tsx
@@ -31,7 +31,7 @@ export default function SignUpForm({
 				},
 				{
 					onSuccess: () => {
-						router.push("/dashboard");
+						router.push("/");
 						toast.success("Sign up successful");
 					},
 					onError: (error) => {
