+++ apps/web/src/lib/api-client.ts
@@ -1,0 +1,17 @@
+import axios from "axios";
+
+export const apiClient = axios.create({
+  baseURL: process.env.NEXT_PUBLIC_SERVER_URL || "http://localhost:3000",
+  headers: {
+    "Content-Type": "application/json",
+  },
+});
+
+// Optional: interceptors for logging/errors
+// apiClient.interceptors.response.use(
+//   (response) => response,
+//   (error) => {
+//     console.error("API Error:", error);
+//     return Promise.reject(error);
+//   }
+// );
